requirements.txt
# requirements.txt

# Web framework
Flask==2.2.5
Flask-Caching==2.0.2

# Data processing
pandas==1.5.3
numpy==1.24.3
nltk==3.8.1

# API and HTTP
requests==2.28.2

# Geolocation
geopy==2.3.0
geocoder==1.38.1

# Configuration
python-dotenv==1.0.0

# Testing
pytest==7.3.1

# Deployment
gunicorn==20.1.0